% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeedMaker-package.R
\docType{package}
\name{SeedMaker-package}
\alias{SeedMaker}
\alias{SeedMaker-package}
\title{SeedMaker: Generate a Collection of Seeds from a Single Seed}
\description{
A mechanism for easily generating and organizing a collection of seeds from a single seed, which may be subsequently used to ensure reproducibility in processes/pipelines that utilize multiple random components (e.g., trial simulation).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hdhshowalter/SeedMaker}
  \item Report bugs at \url{https://github.com/hdhshowalter/SeedMaker/issues}
}

}
\author{
\strong{Maintainer}: Hollins Showalter \email{hollins.showalter@gmail.com}

Other contributors:
\itemize{
  \item  Eli Lilly and Company [copyright holder, funder]
}

}
\keyword{internal}
